Create MutualSubstitution.cpp
A mutual letter substitution is a pair of letters subFirst, subSecond and denotes an operation where letter subSecond is substituted in the place of every occurrence of letter subFirst, and vice versa. Given a string s of n lowercase letters and mutual substitution pair of letters: firstSub, secondSub. The goal is to perform all the substitutions and return the resulting string.